\name{gene_example}
\alias{gene_example}
\docType{data}
\title{
Gene example data
}
\description{
Dummy genes to test data with.
}
\usage{data(gene_example)}
\format{
  The format is:
 int [1:15, 1] 111 111 111 222 222 222 222 222 333 333 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr "x"
}
\examples{
data(gene_example)
}
\keyword{datasets}
