\name{GSAgm-package}
\alias{GSAgm-package}
\alias{GSAgm}
\docType{package}
\title{GSAgm package}
\description{
  This package runs gene set analyses on SNP and RNA data.  See
  individual functions for details. 
}
\details{
\tabular{ll}{
Package: \tab GSAgm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-04\cr
License: \tab GPL-2\cr
}
}
\author{
Rama Raghavan, Alice Wang
Maintainer: Rama Raghavan <rraghavan@kumc.edu>
}
\references{
  1.	Biernacka JM, Jenkins GD, Wang L, et al. Use of the gamma method
  for self-contained gene-set analysis of SNP data. \emph{Eur J Hum
    Genet} 2012;20(5):565-71.
  
2.	Fridley BL, Jenkins GD, Grill DE, et al. Soft truncation
thresholding for gene set analysis of RNA-seq data: application to a
vaccine study. \emph{Sci Rep} 2013;3:2898.

3.	Robinson MD, McCarthy DJ, Smyth GK. edgeR: a Bioconductor
package for differential expression analysis of digital gene expression
data. \emph{Bioinformatics} 2010;26(1):139-40.

4.	Zaykin DV, Zhivotovsky LA, Czika W, et al. Combining p-values in
large-scale genomics experiments. \emph{Pharm Stat} 2007;6(3):217-26.

5.      Fridley, B.L., and Biernacka, J.M. (2011). Gene set analysis of SNP
  data: benefits, challenges, and future directions. \emph{Eur J Hum Genet} 19,
  837-843.

}

\keyword{ package }
