\name{rnaseq_counts}
\alias{rnaseq_counts}
\docType{data}
\title{
RNA Seq Test data
}
\description{
RNA test data for example
}
\usage{data(rnaseq_counts)}
\format{
  A data frame with 150 observations on the following 15 variables.
  \describe{
    \item{\code{rnaseqcount1}}{a numeric vector}
    \item{\code{rnaseqcount2}}{a numeric vector}
    \item{\code{rnaseqcount3}}{a numeric vector}
    \item{\code{rnaseqcount4}}{a numeric vector}
    \item{\code{rnaseqcount5}}{a numeric vector}
    \item{\code{rnaseqcount6}}{a numeric vector}
    \item{\code{rnaseqcount7}}{a numeric vector}
    \item{\code{rnaseqcount8}}{a numeric vector}
    \item{\code{rnaseqcount9}}{a numeric vector}
    \item{\code{rnaseqcount10}}{a numeric vector}
    \item{\code{rnaseqcount11}}{a numeric vector}
    \item{\code{rnaseqcount12}}{a numeric vector}
    \item{\code{rnaseqcount13}}{a numeric vector}
    \item{\code{rnaseqcount14}}{a numeric vector}
    \item{\code{rnaseqcount15}}{a numeric vector}
  }
}
\examples{
data(rnaseq_counts)
## maybe str(rnaseq_counts) ; plot(rnaseq_counts) ...
}
\keyword{datasets}
